Adds necessary constraints to DB fields
   - This adds constraints on db fields using a migration
     and tests for them.
   - This patch also moves migrations to their own files.
